Data source url build

+++ app/src/main/java/com/myprojects/msa_project/domain/Shop.kt
@@ -4,7 +4,7 @@ data class Shop(
     val id: String,
     val name: String,
     val imageUrl: String,
-    val distance: Double,
+    val distance: Int,
     val closedBucket: String,
     val address: String
 )
